Combine UIDConnectivityBase and UIDConnectivity in same file.
